+++ processor/src/main/java/nl/xillio/xill/plugins/testing/constructs/IsFalseConstruct.java
@@ -19,7 +19,7 @@ public class IsFalseConstruct extends Construct {
         return new ConstructProcessor(
                 this::process,
                 new Argument("value"),
-                new Argument("message", NULL, ATOMIC)
+                new Argument("message", NULL)
         );
     }