+++ CHANGELOG.md
@@ -2,8 +2,12 @@
 All notable changes to this project will be documented in this file
 
 ## [3.3.0] - unreleased
+### Add
+- Hash.toSHA1 and Hash.toSHA256 constructs [CTC-1332]
+
 ### Change
 - Add a new stream api for data streaming [CTC-1373]
+- Refactory Encode.toBase64, Decode.fromBase6, Hash.toMD5 to work with streams [CTC-1332]
 
 ## [3.2.0] - 11-02-2016
 ### Add